route cleanup: notifications
